Moving methods from trajectory_accessor.py to either traj.py, traj1d.py or traj2d.py
